move Marker out of MapView component folder since it is its own component
