+++ packages/apollos-ui-mapview/src/Marker.js
